Move link fixes to a plugin file
